+++ src/memory_zone.c
@@ -22,7 +22,7 @@ void    show_alloc_mem(void)
                 block_start = (void *)((char *)aux_block + BLOCK_OFFSET);
                 block_end = (void *)((char *)aux_block + BLOCK_OFFSET + aux_block->size);
 
-                ft_printf("%p - %p : %d bytes\n", block_start, block_end, aux_block->size);
+                ft_printf("%p - %p : %u bytes\n", block_start, block_end, (unsigned int)aux_block->size);
                 total_bytes = total_bytes + aux_block->size;
                 aux_block = aux_block->next;
             }
@@ -44,7 +44,7 @@ void    show_alloc_mem(void)
                 block_start = (void *)((char *)aux_block + BLOCK_OFFSET);
                 block_end = (void *)((char *)aux_block + BLOCK_OFFSET + aux_block->size);
 
-                ft_printf("%p - %p : %d bytes\n", block_start, block_end, aux_block->size);
+                ft_printf("%p - %p : %u bytes\n", block_start, block_end, (unsigned int)aux_block->size);
                 total_bytes = total_bytes + aux_block->size;
                 aux_block = aux_block->next;
             }
@@ -66,12 +66,12 @@ void    show_alloc_mem(void)
                 block_start = (void *)((char *)aux_block + BLOCK_OFFSET);
                 block_end = (void *)((char *)aux_block + BLOCK_OFFSET + aux_block->size);
 
-                ft_printf("%p - %p : %d bytes\n", block_start, block_end, aux_block->size);
+                ft_printf("%p - %p : %u bytes\n", block_start, block_end, (unsigned int)aux_block->size);
                 total_bytes = total_bytes + aux_block->size;
                 aux_block = aux_block->next;
             }
             aux_zone = aux_zone->next;
         }
-        ft_printf("Total : %d bytes\n", total_bytes);
+        ft_printf("Total : %u bytes\n", (unsigned int)total_bytes);
     }
 }
